Combine tests and sources
