+++ iora/src/asset_index.rs
@@ -13,12 +13,6 @@ pub enum ListAssetsError {
     BadQuery { query: String, details: String },
 }
 
-#[derive(Error, Debug)]
-pub enum ListAssetsCacheError {
-    #[error("Something went wrong in the cache.")]
-    StorageError,
-}
-
 pub trait AssetIndex {
     fn list_assets(&self, query: &AssetQuery) -> Result<Vec<AssetDescriptor>, ListAssetsError>;
 }
@@ -66,5 +60,46 @@ where
             }
             Err(e) => Err(e),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        AssetDescriptor, AssetIndex, AssetQuery, CachingAssetIndex, MemoryAssetIndex,
+        MemoryAssetIndexCache, NameConstraint, SemVer,
+    };
+    use std::str::FromStr;
+    use std::time::Duration;
+
+    #[test]
+    fn list() {
+        let cache = Box::new(MemoryAssetIndexCache::new(Duration::from_secs(1)));
+        let remote = Box::new(MemoryAssetIndex::default());
+        remote.descriptors.borrow_mut().push(AssetDescriptor::new(
+            "asset_name",
+            &SemVer::from_str("2.45.6").unwrap(),
+            "asset_hash",
+        ));
+        remote.descriptors.borrow_mut().push(AssetDescriptor::new(
+            "asset_name",
+            &SemVer::from_str("3.45.6").unwrap(),
+            "asset_hash",
+        ));
+        remote.descriptors.borrow_mut().push(AssetDescriptor::new(
+            "asset_name",
+            &SemVer::from_str("3.45.7").unwrap(),
+            "asset_hash",
+        ));
+        let catalog = CachingAssetIndex::new(cache, remote);
+
+        let result = catalog
+            .list_assets(&AssetQuery::from((
+                &NameConstraint::Contains("asset".to_string()),
+                &None,
+            )))
+            .expect("t");
+        assert_eq!(result.len(), 1);
+        assert_eq!(result[0].version, SemVer::from_str("3.45.7").unwrap());
     }
 }
