--- iora/tests/tests.rs
@@ -1,5 +1,0 @@
-extern crate iora;
-
-mod asset_descriptor;
-mod constraints;
-mod semver;
